% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic_do_modelo.R
\name{aic_do_modelo}
\alias{aic_do_modelo}
\title{Criterio de Informação de Akaike (AIC)}
\usage{
aic_do_modelo(
  dados,
  variavel_resposta,
  variaveis_preditoras,
  penalizacao = 2,
  bayesiano = FALSE
)
}
\arguments{
\item{dados}{Banco de dados para análise;}

\item{variavel_resposta}{Variável resposta, da qual queremos fazer a predição;}

\item{variaveis_preditoras}{Vetor composto pelas variáveis que serão usadas na predição;}

\item{penalizacao}{Valor de penalização a ser aplicado no AIC (padrão é 2);}

\item{bayesiano}{Lógico que indica se deve ser usada a penalização do BIC (padrão é FALSE).}
}
\value{
Retorna o valor do Critério de Informação de Akaike (AIC) com base na variável resposta e nas variáveis preditoras do modelo.
}
\description{
Calcula o Critério de Informação de Akaike (AIC) para um modelo estatístico com base na variável resposta e nas variáveis preditoras.
\deqn{AIC = -2*\text{log }L(\hat{\beta}|\textbf{X}) + k\cdot p}
onde:
\itemize{
\item \eqn{p} é a quantidade de variáveis explicativas (preditoras) usadas no modelo;
\item \eqn{k} é a penalização no cálculo do Critério;
\item \eqn{\hat{L}} é o valor máximo da função de verossimilhança, obtido pelo EMV \eqn{\hat{\beta}}.
}
}
\details{
A função oferece a opção de usar o Critério de Informação Bayesiano (BIC), ajustando o valor de penalização com base no logaritmo do número de observações.
}
\examples{
aic_do_modelo(proj2::dados, "y", c("x1", "x2"))
aic_do_modelo(proj2::dados, "y", c("x1", "x2"), bayesiano = TRUE)

}
