% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betas.R
\name{betas}
\alias{betas}
\title{betas}
\usage{
betas(x, y)
}
\arguments{
\item{x}{matrix. Matriz com as variáveis explicativas (preditoras).}

\item{y}{matrix. Matriz com a variável resposta.}
}
\value{
Um vetor com os coeficientes da regressão (betas).

Retorna uma matriz com os coeficientes da regressão.
}
\description{
Calcula os coeficientes (betas) da regressão linear utilizando a fórmula: β̂ = (X'X)^(-1)X'Y.
A primeira coluna da matriz x representa o intercepto (adicionado automaticamente).
}
\details{
A função calcula os coeficientes (betas) da regressão linear usando a fórmula
clássica de mínimos quadrados:

\deqn{\beta = (X^T X)^{-1} X^T Y}

Se a matriz \(X^T X\) for singular, a regularização Ridge será aplicada para
garantir a estabilidade dos coeficientes.
}
\examples{
betas(X, Y)
}
