% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{categoricas}
\alias{categoricas}
\title{Exemplo com variáveis categóricas}
\format{
An object of class \code{data.frame} with 78 rows and 4 columns.
}
\usage{
categoricas
}
\description{
Conjunto de dados contendo variáveis numéricas \emph{"y"} e \emph{"numerica"} e variáveis categóricas \emph{"fator1"} e \emph{"fator2"} para
testar as funções do pacote.
}
\details{
O conjunto de dados foi gerado através de simulação, em que as variáveis categóricas foram
obtidas com uso da função \emph{sample} e a variável "numerica" foi dada amostrando de uma Poisson deslocada.
Também foi amostrado um erro normal ao produzir a variável "y" por uma equação.
}
\examples{
data(categoricas)
}
\keyword{datasets}
