% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dados}
\alias{dados}
\title{Exemplo para aplicar funções}
\format{
An object of class \code{data.frame} with 98 rows and 3 columns.
}
\usage{
dados
}
\description{
Conjunto de dados contendo variáveis numéricas "y", "x1" e "x2" para testar as funções do pacote.
}
\details{
O conjunto de dados foi gerado através de simulação, em que as variáveis numéricas foram obtidas
através de amostras da distirbuição normal com diferentes parâmetros. Também foi amostrado um erro
normal ao produzir a variável "y" por uma equação.
}
\examples{
data(dados)
}
\keyword{datasets}
