% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficopo.R
\name{grafico_p_o}
\alias{grafico_p_o}
\title{Gráfico Observados vs Preditos}
\usage{
grafico_p_o(resultado_regressao)
}
\arguments{
\item{resultado_regressao}{list. O resultado da função regressao().}
}
\value{
Um gráfico comparativo entre os valores preditos e observados.

Um gráfico de comparação de valores preditos vs observados, exibido na tela.
}
\description{
Gera um gráfico comparando os valores preditos e os observados dentro do conjunto de dados de treino.
}
\details{
A função plota os valores preditos contra os valores observados para
fornecer uma visão visual da qualidade do ajuste do modelo.
}
\examples{
grafico(X, Y)
}
