% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico.R
\name{grafico_residuos}
\alias{grafico_residuos}
\title{Gráfico Resíduos vs Preditos}
\usage{
grafico_residuos(resultado_regressao)
}
\arguments{
\item{resultado_regressao}{list. O resultado da função regressao().}
}
\value{
Um gráfico de dispersão comparando os valores preditos e os resíduos dentro do conjunto de dados.

Um gráfico de dispersão exibido na tela.
}
\description{
Gera um gráfico de dispersão comparando os valores preditos e os resíduos dentro do conjunto de dados de treino.
}
\details{
A função plota os valores preditos contra os valores observados para
fornecer uma visão visual da qualidade do ajuste do modelo.
}
\examples{
grafico(X, Y)
}
