% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_matrizes.R
\name{modelo}
\alias{modelo}
\title{Modelo}
\usage{
modelo(dados, formula)
}
\arguments{
\item{dados}{data.frame. O banco de dados contendo as variáveis.}

\item{formula}{fórmula. Fórmula no formato Y ~ x1 + x2 + ... + xn.}
}
\value{
Uma lista contendo as matrizes das variáveis preditoras e resposta.

Uma lista contendo a matriz X (explicativas) e a matriz Y (resposta).
}
\description{
Recebe um banco de dados e uma fórmula e retorna duas matrizes:
uma matriz X com as variáveis explicativas e uma matriz Y com a variável resposta,
preparadas para serem utilizadas na função betas.
}
\details{
A função transforma os dados de entrada em matrizes para facilitar o cálculo
da regressão, separando as variáveis preditoras e resposta.
}
\examples{
extrair_matrizes(dados, Crescimento_PIB ~ Inflacao + Gastos_Governamentais)

}
