% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosso_lm.R
\name{nosso_lm}
\alias{nosso_lm}
\title{Regressão Linear Múltipla}
\usage{
nosso_lm(dados, variavel_resposta, variaveis_preditoras)
}
\arguments{
\item{dados}{Banco de dados para análise;}

\item{variavel_resposta}{Variável resposta, da qual queremos
fazer a predição;}

\item{variaveis_preditoras}{Vetor composto pelas variáveis
que serão usadas na predição.}
}
\value{
Retorna-se um objeto do tipo "nosso_lm", contendo uma lista de informações
que podem ser obtidas ao estabeler o modelo. Nesta lista, encontra-se:
\itemize{
\item .$betas: Relação com estimativas para os coeficientes da regressão;
\item .$qme: Quadrado médio do erro, estimador não viesado para variância do erro normal;
\item .$r2_ajustado: Coeficiente de determinação ajustado pela quantidade de preditoras;
\item .$r2: Coeficiente de determinação sem ajuste;
\item .$residuos: Resíduos obtidos pela diferença entre o que foi observado e o que foi predito no modelo;
\item .$semi_studentizado: Resíduos semi_studentizados;
\item .$studentizado: Resíduos studentizados;
\item .$valores_preditos: Predição dos valores da variável resposta em relação ao modelo construído;
\item .$aic_do_modelo: Critério de informação de Akaike para o modelo, com \eqn{k=2};
\item .$matriz_delineamento: Matriz com informações sobre as variáveis preditoras;
\item .$variavel_resposta: Lista contendo nome da variável resposta e seus valores.
}
}
\description{
Retorna uma lista com diversas informações relacionadas a um modelo de
regressão linear múltipla, construído com base em um banco de dados
e suas variáveis. O modelo é o que segue.

\deqn{Y=\beta_{0}+\beta_{1}X_{1}+\cdots+\beta_{p-1}X_{p-1}+\epsilon}

Temos, para a equação acima,
\itemize{
\item \eqn{Y} é a variável resposta;
\item \eqn{X_{1},...,X_{p-1}} são as variáveis preditoras;
\item \eqn{\beta_{0},\beta_{1},...,\beta_{n}} são os coeficientes da regressão; \eqn{\beta_{0}} é o intercepto;
\item \eqn{\epsilon\sim\mathcal{N}(0,\sigma^{2})} é o erro aleatório independente e identicamente distribuído para toda observação.
}

Além disso, permite utilização de uma série de métodos:
\itemize{
\item print;
\item plot;
\item confint;
\item summary;
\item predict.
}
}
\details{
É necessário que a matriz de delineamento tenha posto completo.

Verifique se os argumentos da função estão corretamente inseridos. É
necessário que as variáveis estejam entre "aspas" (ou seja, char com
nome da variável).

Para uso do módulo \emph{predict}, é preciso especificar, em valores numéricos,
os valores da variáveis preditoras na ordem em que aparecem na equação de regressão.
Deste modo, por exemplo, deve ter a inserção correta de 0s e 1s nas variáveis dummy.
}
\examples{
regressao <- nosso_lm(proj2::dados, "y", c("x1", "x2"))
plot(regressao)
confint(regressao)
summary(regressao)
predict(regressao, observacao = c(13.82, 62.17))

regressao <- nosso_lm(proj2::categoricas, "y", c("fator1", "fator2", "numerica"))
plot(regressao)
confint(regressao)
summary(regressao)
predict(regressao, observacao = c("b", "c", 5)) # vai dar errado
predict(regressao, observacao = c(1,0,0,0,1,0,5)) # vai dar certo: segue a relação dos betas

}
