% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_valores.R
\name{p_valores}
\alias{p_valores}
\title{P-valores para cada beta}
\usage{
p_valores(x, y)
}
\arguments{
\item{x}{matrix. Matriz com as variáveis explicativas.}

\item{y}{matrix. Matriz com a variável resposta.}
}
\value{
Um vetor com os p-valores associados a cada coeficiente.

Uma lista com os betas e seus p-valores.
}
\description{
Calcula o p-valor para cada coeficiente da regressão (beta), assumindo que o erro segue uma distribuição normal.
Retorna uma lista dos coeficientes com seus p-valores e um asterisco ao lado se o p-valor for menor que 0.05.
}
\details{
A função utiliza os coeficientes e a matriz de covariância para calcular os
p-valores com base na distribuição t de Student.
}
\examples{
p_valores(betas, cov_matrix)
}
