% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peixes.R
\docType{data}
\name{peixes}
\alias{peixes}
\title{Conjunto de dados do pacote}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 159 rows and 7 columns.
}
\usage{
peixes
}
\description{
lista de peixe!
}
\keyword{datasets}
