% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicao.R
\name{predicao}
\alias{predicao}
\title{Predição}
\usage{
predicao(modelo_retorno, z)
}
\arguments{
\item{modelo_retorno}{list. O retorno da função modelo(), contendo as matrizes X e Y.}

\item{z}{matrix. Matriz com os novos valores das variáveis explicativas (sem o intercepto).}

\item{modelo}{Um modelo de regressão ajustado.}

\item{dados}{O conjunto de dados utilizado para realizar as predições.}
}
\value{
Um vetor numérico com os valores preditos.

Retorna um data frame com as variáveis preditoras e os valores preditos.
}
\description{
Calcula os valores preditos para um novo conjunto de dados utilizando os coeficientes estimados no conjunto de treino.
}
\details{
Esta função usa os coeficientes estimados pelo modelo de regressão e aplica-os
às variáveis preditoras para calcular os valores preditos de Y.
Se houver multicolinearidade, a regularização Ridge é aplicada.
}
\examples{
modelo_resultado <- modelo(dados, Crescimento_PIB ~ Inflacao + Gastos_Governamentais)
predicao(modelo_resultado, dados)
}
