% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predito_observado.R
\name{predito_observado}
\alias{predito_observado}
\title{Gráfico: Valores Preditos v.s. Valores Observados}
\usage{
predito_observado(dados, variavel_resposta, variaveis_preditoras, reta = TRUE)
}
\arguments{
\item{dados}{Banco de dados para análise;}

\item{variavel_resposta}{Variável resposta, da qual queremos
fazer a predição;}

\item{variaveis_preditoras}{Vetor composto pelas variáveis
que serão usadas na predição;}

\item{reta}{Lógico que adiciona a reta identidade y = x (padrão é TRUE).}
}
\value{
Retorna-se um gráfico de dispersão das variáveis respostas preditas no modelo v.s. observadas.
}
\description{
Com ajuda do pacote \emph{ggplot2}, cria um gráfico de dispersão da variável resposta predita v.s. observada com a reta identidade para comparar os valores.
}
\details{
Verifique se os argumentos da função estão corretamente inseridos. É necessário que as variáveis estejam entre "aspas" (ou seja, char com nome da variável).
}
\examples{
predito_observado(proj2::dados, "y", c("x1", "x2"))

}
