% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preditos.R
\name{preditos}
\alias{preditos}
\title{Preditos}
\usage{
preditos(x, y)
}
\arguments{
\item{x}{matrix. Matriz com as variáveis explicativas de treino.}

\item{y}{matrix. Matriz com a variável resposta.}
}
\value{
Um vetor com os valores ajustados (ou preditos) do modelo.

Retorna um vetor com os valores preditos.
}
\description{
Calcula os valores preditos dentro do próprio conjunto de dados de treino.
}
\details{
A função utiliza os coeficientes estimados pela função betas() e aplica-os à
matriz X para prever os valores de Y. O resultado é um vetor com os valores preditos.
}
\examples{
preditos(X, Y)
}
