% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qme.r
\name{qme}
\alias{qme}
\title{Quadrado Médio do Erro}
\usage{
qme(dados, variavel_resposta, variaveis_preditoras)
}
\arguments{
\item{dados}{Banco de dados para análise;}

\item{variavel_resposta}{Variável resposta, da qual queremos
fazer a predição;}

\item{variaveis_preditoras}{Vetor composto pelas variáveis
que serão usadas na predição.}
}
\value{
Retorna-se o valor do Quadrado Médio do Erro.
}
\description{
Encontra o Quadrado Médio do Erro (QME)
\deqn{QME = \dfrac{1}{n-p}\sum_{i=1}^{n}{(\hat{y}_i - y_i)^2}}
Temos, para a equação acima,
\itemize{
\item \eqn{\hat{y}_i} é a estimativa da variável resposta na observação i;
\item \eqn{y_i} é o valor da observação i da variável resposta;
\item \eqn{n} é o número de observações do conjunto de dados;
\item \eqn{p} é a quantidade de preditoras utilizadas na regressão, incluindo intercepto.
}
}
\details{
Verifique se os argumentos da função estão corretamente inseridos. É necessário que as variáveis estejam entre "aspas" (ou seja, char com nome da variável).
}
\examples{
qme(regressaolinear::dados, "y", c("x1", "x2"))

}
