% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_residuos.r
\name{qq_residuos}
\alias{qq_residuos}
\title{QQ-Plot dos resíduos}
\usage{
qq_residuos(dados, variavel_resposta, variaveis_preditoras, reta = TRUE)
}
\arguments{
\item{dados}{Banco de dados para análise;}

\item{variavel_resposta}{Variável resposta, da qual queremos
fazer a predição;}

\item{variaveis_preditoras}{Vetor composto pelas variáveis
que serão usadas na predição;}

\item{reta}{Lógico que adiciona a reta do esperado sob distribuição normal (padrão é TRUE).}
}
\value{
Retorna-se um gráfico QQ-plot dos resíduos do modelo de regressão.
}
\description{
Com ajuda do pacote \emph{ggplot2}, cria um gráfico QQ-Plot dos resíduos do modelo de regressão.
}
\details{
Verifique se os argumentos da função estão corretamente inseridos. É necessário que as variáveis estejam entre "aspas" (ou seja, char com nome da variável).
}
\examples{
qq_residuos(proj2::dados, "y", c("x1", "x2"))

}
