% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{r2}
\alias{r2}
\title{Coeficiente de determinação (\eqn{R^2})}
\usage{
r2(dados, variavel_resposta, variaveis_preditoras)
}
\arguments{
\item{dados}{Banco de dados para análise;}

\item{variavel_resposta}{Variável resposta, da qual queremos
fazer a predição;}

\item{variaveis_preditoras}{Vetor composto pelas variáveis
que serão usadas na predição.}
}
\value{
Retorna-se o coeficiente de determinação (\eqn{R^2}) com base na variável resposta e variáveis preditoras do modelo.
}
\description{
Calcula o coeficiente de determinação do modelo com base na variável resposta e nas variáveis preditoras.
\deqn{R^2 = 1 - \dfrac{SQ_{res}}{SQ_{tot}}}
Temos, para a equação acima,
\itemize{
\item \eqn{SQ_{res}} é a soma dos quadrados dos resíduos do modelo;
\item \eqn{SQ_{tot}} é a soma dos quadrados total do modelo, relativa à variável resposta.
}
}
\details{
Verifique se os argumentos da função estão corretamente inseridos. É necessário que as variáveis estejam entre "aspas" (ou seja, char com nome da variável).
}
\examples{
r2(proj2::dados, "y", c("x1", "x2"))

}
