% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2_ajustado.R
\name{r2_ajustado}
\alias{r2_ajustado}
\title{Coeficiente de determinação ajustado (\eqn{R_{aj}^2})}
\usage{
r2_ajustado(dados, variavel_resposta, variaveis_preditoras)
}
\arguments{
\item{dados}{Banco de dados para análise;}

\item{variavel_resposta}{Variável resposta, da qual queremos
fazer a predição;}

\item{variaveis_preditoras}{Vetor composto pelas variáveis
que serão usadas na predição.}
}
\value{
Retorna-se o coeficiente de determinação ajustado (\eqn{R_{aj}^2}) com base na variável resposta e variáveis preditoras do modelo.
}
\description{
Calcula o coeficiente de determinação ajustado do modelo com base na variável resposta e nas variáveis preditoras.
\deqn{R_{aj}^2 = 1 - \dfrac{n - 1}{n - p}\dfrac{SQ_{res}}{SQ_{tot}}}
Temos, para a equação acima,
\itemize{
\item \eqn{n} é quantidade de observações no conjunto de dados;
\item \eqn{p} é a quantidade de variáveis explicativas (preditoras) usadas no modelo, mais o intercepto;
\item \eqn{SQE} é soma quadrática dos resíduos;
\item \eqn{SQT} é soma quadrática total, relativa à variável resposta.
}
}
\details{
Verifique se os argumentos da função estão corretamente inseridos. É necessário que as variáveis estejam entre "aspas" (ou seja, char com nome da variável).
}
\examples{
r2_ajustado(proj2::dados, "y", c("x1", "x2"))

}
