% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressao.R
\name{regressao}
\alias{regressao}
\title{Regressão}
\usage{
regressao(modelo_retorno, R = 0)
}
\arguments{
\item{modelo_retorno}{list. O retorno da função modelo(), contendo as matrizes X e Y.}

\item{R}{numeric. Parâmetro opcional para indicar se os resultados devem ser impressos (R = 1).}
}
\value{
Uma lista contendo os coeficientes da regressão, p-valores, resíduos, R² e R² ajustado.

Uma lista contendo os coeficientes (betas), os p-valores para cada beta, e os resíduos da regressão.
}
\description{
Aplica as funções betas, p_valores e residuos às matrizes X e Y fornecidas e retorna os resultados das três funções.
Se o parâmetro R for igual a 1, os resultados serão impressos na tela.
}
\details{
A função realiza a regressão linear, ajusta o modelo e, em caso de multicolinearidade,
utiliza regularização Ridge para estabilizar o cálculo. Retorna os principais indicadores
como betas, p-valores, resíduos, R² e R² ajustado.
}
\examples{
modelo_resultado <- regressao(dados, Crescimento_PIB ~ Inflacao + Gastos_Governamentais)
}
