% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos.R
\name{residuos}
\alias{residuos}
\title{Resíduos}
\usage{
residuos(x, y)
}
\arguments{
\item{x}{matrix. Matriz com as variáveis explicativas.}

\item{y}{matrix. Matriz com a variável resposta.}
}
\value{
Um vetor com os resíduos, que são a diferença entre os valores observados e os preditos.

Retorna um vetor com os resíduos.
}
\description{
Calcula os resíduos da regressão, que são definidos como a diferença entre os valores observados e os preditos.
}
\details{
A função calcula os resíduos, que são a diferença entre os valores observados e
os valores preditos pelo modelo de regressão.
}
\examples{
residuos(X, Y)
}
