% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studentizado.R
\name{studentizado}
\alias{studentizado}
\title{Resíduos Studentizados}
\usage{
studentizado(dados, variavel_resposta, variaveis_preditoras)
}
\arguments{
\item{dados}{Banco de dados para análise;}

\item{variavel_resposta}{Variável resposta, da qual queremos analisar os resíduos;}

\item{variaveis_preditoras}{Vetor composto pelas variáveis que foram utilizadas na predição.}
}
\value{
Retorna um vetor contendo os resíduos studentizados.
}
\description{
Calcula os resíduos studentizados de um modelo de regressão com base na variável resposta e nas variáveis preditoras.
A fórmula utilizada é:
\deqn{r_i^* = \dfrac{e_i}{\sqrt{QME \cdot (1 - h_i)}}}
onde:
\itemize{
\item \eqn{e_i} é o resíduo da observação \eqn{i};
\item \eqn{QME} é o quadrado médio do erro do modelo;
\item \eqn{h_i} é a alavancagem da observação \eqn{i}.
}
}
\details{
Certifique-se de que as variáveis estão corretamente especificadas e que os dados não contêm valores ausentes.
}
\examples{
studentizado(proj2::dados, "y", c("x1", "x2"))

}
