% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valores_preditos.R
\name{valores_preditos}
\alias{valores_preditos}
\title{Valores preditos para a variável resposta}
\usage{
valores_preditos(dados, variavel_resposta, variaveis_preditoras)
}
\arguments{
\item{dados}{Banco de dados para análise;}

\item{variavel_resposta}{Variável resposta, da qual queremos
fazer a predição;}

\item{variaveis_preditoras}{Vetor composto pelas variáveis
que serão usadas na predição.}
}
\value{
Retorna-se a relação dos valores preditos na mesma ordem que estão as observações.
}
\description{
Encontra os valores preditos na regressão
linear múltipla após obter o modelo,
\deqn{\hat{Y}=X\hat{\beta}}
Temos, para a equação acima,
\itemize{
\item \eqn{\hat{Y}} é o vetor com variáveis respostas preditas de acordo com o modelo;
\item \eqn{X} é a matriz de delineamento;
\item \eqn{\hat{\beta}} é o vetor contendo coeficientes estimados.
}
}
\details{
É necessário que a matriz de delineamento tenha posto completo.

Verifique se os argumentos da função estão corretamente inseridos. É necessário que as variáveis estejam entre "aspas" (ou seja, char com nome da variável).
}
