% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betas.R
\name{betas}
\alias{betas}
\title{Coeficientes da Regressão}
\usage{
betas(dados, variavel_resposta, variaveis_preditoras)
}
\arguments{
\item{dados}{Banco de dados para análise;}

\item{variavel_resposta}{Variável resposta, da qual queremos
fazer a predição;}

\item{variaveis_preditoras}{Vetor composto pelas variáveis
que serão usadas na predição.}
}
\value{
Relação dos coeficientes e suas estimativas.
}
\description{
Encontra os coeficientes na regressão
linear múltipla,
\deqn{Y = \beta_{0}+\beta_{1}X_{1}+...+\beta_{p-1}X_{p-1}+\epsilon}.
}
\details{
É necessário que a matriz de delineamento tenha posto completo.
}
