% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_residuos.R
\name{grafico_residuos}
\alias{grafico_residuos}
\title{Gráfico: Valores Preditos v.s. Resíduos}
\usage{
grafico_residuos(
  dados,
  variavel_resposta,
  variaveis_preditoras,
  studentizado = "nao"
)
}
\arguments{
\item{dados}{Banco de dados para análise;}

\item{variavel_resposta}{Variável resposta, da qual queremos
fazer a predição;}

\item{variaveis_preditoras}{Vetor composto pelas variáveis
que serão usadas na predição;}

\item{studentizado}{Parâmetro para indicar se os resíduos são calculados sem normalização (argumento "nao"),
se serão semi-studentizados (argumento "semi") ou studentizados (argumento "sim").}
}
\value{
Retorna-se um gráfico de dispersão das variáveis preditas v.s. observadas.
}
\description{
Com ajuda do pacote \emph{ggplot2}, cria um valores preditos para a variável resposta v.s. resíduos, podendo ou não ser studentizados.
}
\details{
Verifique se os argumentos da função estão corretamente inseridos. É necessário que as variáveis estejam entre "aspas" (ou seja, char com nome da variável).
}
\examples{
grafico_residuos(proj2::dados, "y", c("x1", "x2"), studentizado = "sim")

}
