% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicao.R
\name{predicao}
\alias{predicao}
\title{Predição através do modelo de regressão}
\usage{
predicao(observacao, dados, variavel_resposta, variaveis_preditoras)
}
\arguments{
\item{observacao}{Novos valores para variáveis preditoras, dos quais queremos obter predição;}

\item{dados}{Banco de dados para análise;}

\item{variavel_resposta}{Variável resposta, da qual queremos
fazer a predição;}

\item{variaveis_preditoras}{Vetor composto pelas variáveis
que serão usadas na predição.}
}
\value{
Retorna-se a relação dos valores preditos na mesma ordem que estão as observações.
}
\description{

}
\details{
Verifique se os novos dados estão na ordem em que foram colocados os betas da predição, no argumento variavel_preditoras.

É necessário que a matriz de delineamento tenha posto completo.

Verifique se os argumentos da função estão corretamente inseridos. É necessário que as variáveis estejam entre "aspas" (ou seja, char com nome da variável).
}
