% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos.R
\name{residuos}
\alias{residuos}
\title{Resíduos obtidos}
\usage{
residuos(dados, variavel_resposta, variaveis_preditoras)
}
\arguments{
\item{dados}{Banco de dados para análise;}

\item{variavel_resposta}{Variável resposta, da qual queremos
fazer a predição;}

\item{variaveis_preditoras}{Vetor composto pelas variáveis
que serão usadas na predição.}
}
\value{
Retorna-se a relação dos resíduos na mesma ordem que estão as observações.
}
\description{
Encontra os resíduos na regressão
linear múltipla após obter o modelo,
\deqn{e=Y-X\hat{\beta}}
Temos, para a equação acima,
\itemize{
\item \eqn{e} é o resíduo obtido;
\item \eqn{Y} é a vetor variável resposta observado;
\item \eqn{X} é a matriz de delineamento;
\item \eqn{\hat{\beta}} é o vetor contendo coeficientes estimados.
}
}
\details{
É necessário que a matriz de delineamento tenha posto completo.

Verifique se os argumentos da função estão corretamente inseridos. É necessário que as variáveis estejam entre "aspas" (ou seja, char com nome da variável).
}
\examples{
residuos(regressaolinear::dados, "y", c("x1", "x2"))

}
